requirements.txt
fastapi==0.109.0
aiofiles>=23.0.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pinecone-client==3.0.0
pymupdf==1.23.21
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
xlrd>=2.0.1
python-multipart==0.0.9
google-genai>=0.3.0
anthropic>=0.42.0
tenacity>=8.0.0
requests>=2.31.0
passlib[bcrypt,argon2]==1.7.4
python-jose==3.3.0
ezdxf>=1.1.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0